requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
click==7.1.2
    # via uvicorn
faker==6.6.3
    # via -r requirements.in
fastapi==0.63.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
httptools==0.1.1
    # via uvicorn
pydantic==1.8.1
    # via
    #   -r requirements.in
    #   fastapi
python-dateutil==2.8.1
    # via faker
python-dotenv==0.17.0
    # via uvicorn
pyyaml==5.4.1
    # via uvicorn
six==1.15.0
    # via python-dateutil
starlette==0.13.6
    # via fastapi
text-unidecode==1.3
    # via faker
typing-extensions==3.7.4.3
    # via pydantic
uvicorn[standard]==0.13.4
    # via -r requirements.in
uvloop==0.15.2
    # via uvicorn
watchgod==0.7
    # via uvicorn
websockets==8.1
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
